In the last commit we have broke everythin, fix get_timestamp, get_channel_id - we have the data in message and not in message_data. Convert timestamp in iso format with utc
